Code cleanup - fixing /w4 warnings and improving use of integer types - Except HUD.cpp because that will be overhauled soon for other reasons anyway

+++ Source/Dialogs.h
@@ -35,7 +35,7 @@ struct ZERO_INIT(TreeViewDialogItem)
 	int parentIndex;
 	bool isRootCategory;
 };
-size_t layoutItemSelect(const std::vector<TreeViewDialogItem*>& theList);
+int layoutItemSelect(const std::vector<TreeViewDialogItem*>& theList);
 
 void targetAppPath(std::string& thePath, std::string& theCommandLineParams);
